Add a NullableAddress class to support an EmptyAddress type.
Currently, the `add` command does not support an optional address field, tags can be optional since it is implemented in the form of a `Set`.

This change is necessary since our specification requires Address to be an optional field. Introducing the NullableAddress and EmptyAddress class allows the user to avoid specifying the customer's address when calling the `addcus` command.

Since any address input by the user must be non-empty, the EmptyAddress can be stored as the empty string within the JSON storage file without any conflicts with any Address instances, and no changes to the storage representation is necessary.