Model: fix filteredCommissions
Test cases are failing as they may not have customers.

This is a hacky fix for filteredCommissions - not used currently, but
may be necessary when search is used in the future.

When it is used, it will need significant changes, because unlike a
person list that is always the same source list, this is not true for
commissions.