Update Person.java
isSamePerson method: make it check for phone number and email address.

isSamePerson method checks for same name to determine if 2 Customers are the same.

Different Customers can have the same name but are different people.

Let's make isSamePerson check for phone number and email address which are unique to each Customer.

Using unique identifiers instead is preferable because this allows 2 Customers to have same names but both to exist in bobaBot.