Implement Edit Command that supports Email address lookups
Edit Command: updates a Customer's details via index of Customer in underlying implementation of an ArrayList.

This is not practical to keep track in real life as the indexes are constantly changing, especially if we want to delete a Customer from bobaBot.

Identifying a Customer via his or her Email address allows us to bypass this problem while ensuring the identifier is kept unique.

Let's use the Email address to search for the underlying index instead of having the index provided at input.