Update DeleteCommand for v1.2
The current DeleteCommand is executed using index person in the list.

This limits the ability of the cashier to delete/manage customers since
the implementation only supports index.

The new DeleteCommand is no longer executed via index but via 2 options:
* Phone Number or
* Email
The cashier can now delete the customers by specifying either the prefix for
phone number 'p/PHONE_NUMBER' or prefix for email 'e/EMAIL'.

This offers more flexibility in the fields the cashier can use to delete/manage
the customers.