AddCommandParser and AddressBookParser have similar logic. Furthermore, AddressBookParser has a very long method.
Code de-duplication will make it easier to extend for all commands. Simplifying the switch statements will enhance readability and code performance.

Let's create enums to associate each the parse function with each possible command from the user. Let's also make CommandParser an abstract class so that other parsers can benefit from the shared parse function.

Using enums is more performant than using HashMap for a key-value storage, and is easy to modify. The abstract class contains the method parse which is final and is the driving code for the parsing logic. The parse method calls handleParsedCommand, which is implementd by the subclasses.