Change help window
There is a help window with a link to AB3 user guide.

The link should be updated to Mass Linkers user guide.
A command summary could be included for easy reference.
Changing the button from copy URL to open URL could provide
more convenience for users to directly access user guide with a click.

Let's
* update the link to Mass Linkers user guide
* add a command summary
* allow user to open user guide directly in the browser with a button