Improve Find feature to be more versatile
Find command only returns names that fufill full word matching.

A more general find feature matches the general nature of the
word, and is more user-friendly as the user might want to search
for a person based on details other than the name and may only
remember a part of a name or email for example.

Let's,
* update the search algorithm to use non full word matching
* update the search algorithm to search all details of a person