Add FilteredStudents class to improve OOP.
A filtered list of students and type of info to display on the GUI are
encapsulated in the FilteredStudents class.
The PersonListPanel is associated with FilteredStudents instead of Model
as the panel should only have access to the students and type of info
to display.

The viewAll command now will only show the Person's name and tags. Once
the student's ID is implemented, the viewAll command will show the
student's name and ID.