Move Address and PropertyAddress into separate package
PropertyAddress is in the property package as it is a field in Property,
since it would not make sense to make use of the original Address
class in the person package. it is not named "Address" as well, since that
would mean conflicting class names in ParserUtil, where both classes
need to be imported.

however, since both PropertyAddress and Address have the same
implementation, it makes sense to have them as one class.

let's move both classes out to a separate address package. we then delete the
PropertyAddress class, and change all references to refer to the Address
class's new location in the address package instead.