Add command: modify email and phone to become optional parameters, refactor address to remark
Email, phone, and address are mandatory parameters.

Email and phone are not crucial information for an internship application, thus making them
optional parameters is more suitable. When email or phone is not supplied by the users, they
will take the default value which is "NA". Address is refactored to remark because address
is irrelevant for an internship application. Remark is also optional parameters.

Let's,
* make email, phone, and address optional parameters in add command
* refactor address to remark
* improve GUI to display the details in a more readable format
* modify test cases to support the changes in add command
* fix checkstyle errors