Add Birthday field to Person object
Adding Birthday field allows Financial Advisors to keep track of
the birthdays of their contacts.

For now: Add Birthday class which stores a simple String value
to represent the birthday. Person model to have a Birthday field.
Update current test cases to use Person objects with Birthday fields.

To be added: Parse String such that birthdays are displayed in
standardised format. Tests for input validation also to be done.