Add JsonAdaptedReview
The code does not properly handle reviews yet, using JsonAdaptedStall
as a placeholder. As Review has not been added as of this commit, this
would use Stall in place of Review.

Let's add the JsonAdaptedReview class and properly work with the
AddressBook to convert to and from the model and JSON.