Modify dummy values to prevent potential bugs
Previously, all dates 2001-01-01 counted as dummy values for mastery
check and consultation, and all empty strings counted as dummy values
for Telegram handle. This was done to make the Telegram, consultation,
and mastery check fields optional. Dummy values would not be displayed
on FRIDAY, so this was a potential bug.

Now, dummy values (renamed to empty values) are singular instances of
the respective classes, and hence no user inputs will conflict with
empty values. Tests have been added to ensure this behaviour.

However, this caused some bugs with loading in json data from storage,
as the empty instances would be saved with "empty" values, thus causing
them to fail the isValid check. A new method isValidOrEmpty was written
and used in the JsonAdaptedPerson class, whereby empty values would be
correctly replaced with empty instances in the Student object.