Modify parseConsultation and parseMasteryCheck to catch date exceptions
Previously, LocalDateTimeParseExceptions were caught in
AddCommandParser, since the parse methods took in LocalDate objects as
argument. It was an inelegant solution as handling the exceptions in
the parse methods made more sense.

The methods have been modified to construct the LocalDate objects within
them so that they can catch the exceptions more naturally.