UI: add Commissions GUI
Adds the GUI for commissions
- Enable tracking of the 'active' Commission in the `ModelManager`
- Adds `CommissionDetailsPane` to show the details of the selected commission
- Updates GUI to change commission pane to show the 'current' commission when
a commission card is clicked

Implementation note: This implementation would currently reset the selected
commission everytime a customer is selected. This design choice was undertaken
so that when a different customer is selected, the commission detail pane will
not show the previous commission belonging to the previous selected customer.