modify the add command to take in a -p argument
Currently, the add command cannot differentiate between commands to add
a property and to add a client.

We have to change it to be able to differentiate between both, so that
users can choose whether to add a new property or client.

Let's allow the add command to take in '-p' directly after 'add', to
allow users to specify that they would like to add a 'Property'.

Note:
- '-c' for adding clients has not been implemented yet.