Add Role field to an Internship
An Internship should not only have a Company name, but also a Role (e.g
Backend Engineer).

Let's add an attribute to the Internship Model, and make the necessary
changes to all test files as well.

One can think of Company name and Role as a "primary key" or unique
identifier for an Internship in WorkBook. As such, we should allow for
entries that have the same Company as long as they have different Roles.

Take a look at Internship#isSameInternship for this functionality (this
method is used throughout the codebase e.g during Edit, loading from
.json, etc)