Change equality condition of AddcCommand
Two AddcCommand should be equal if and only if the module class names are
the same. The sessions for the specified module classes should not be
compared against each other during the equality check because it does
not make sense to compare sessions of module classes that are not yet
created (they are created via the addc command). It also further
complicates testing.

Let's use ModuleClass#isSameModuleClass instead of ModuleClass#equals to
perform the equality check.