Added Status command: updates the status of the person with the command words: "r", "R" for Rejected, "o", "O" for Offered and "p",  "P" for progress after the s/ prefix.
There might be format inconsistency as the User currently uses
different command words for the add adn edit command for status.

The user is also able to edit the current status with the edit
command which puts the functionality of this command under question,
maybe this can be used as a shortcut for the user to quickly update
his status?