Fix: Wrong parseTitle ParseException
`ParserUtil#parseTitle` was wrongly throwing a `ParseException` with
exception message for `Phone` constraints. As as result, when you
enter an `addcom` command with an invalid title (one that contains
non-alphumerical characters like punctuation), the result displays
"Phone numbers should only contain numbers, and it should be at
least 3 digits long". This commit fixes this bug.