UI: update iteration list GUI
Updates the iteration list GUI in the `CommissionDetailsPane` by

- Fixing the bug whereby the iteration list is not cleared when
no commission is seleted (ie: the iteration list of the previously
selected commission continues to show even after changing customer,
and no commission is presently selected). This bug is due to the
`ListView` not being set to `null` when no commission is selected.

- Improving the GUI by displaying the iteration details more
neatly in the `IterationListItem`. Now, the Iteration index,
`Description`, and `Date` will be aligned in a single row
at the top of the `IterationListItem`. The `Description` field
will also scale according to the window size while the other two
fields remain fixed. A minor problem that has not been fixed yet
is that for very long description names, the description will be
eclipsed (...).

Future improvements should look into scaling this
height-wise to show the full description (or impose a character cap).
Note that this issue is also present for other labels in AB (like
Commission title).

- This commit also enforces a minimum width (equal to the `CustomerListPanel`
or `CommissionListPanel`) to the `CustomerDetailsPane` and `CommissionDetailsPane`
to reduce the pain with having to deal with alignment issues in `IterationListItem`
when resizing the panes.