UI: update `CommissionDetailsPane` UI
Updates the `CommissionDetailsPane` by
- Adding the field icons to the attributes (Deadline, Customer and Fee)
- Makes the CompletionStatus more meaningful by adding a method to
translate that into more 'meaningful' terms ("Completed", "In Progress",
and "Not Started") based on the `CompletionStatus` and number of Iterations
in the Commission.

In order to achieve this, these new methods/ enums were added:
- Enum `CompletionStatusString` which has all the valid CompletionStatusStrings
used to render the more 'meaningful' completion statuses
- `Commission#getCompletionStatusString` which computes the CompletionStatusString
based on its CompletionStatus (if completed, "Completed"), and the number of
iterations (0: "Not Started", >0: "In Progress")
- To support this, a `size` method was also created in `UniqueIterationList`
to retrieve the number of iterations stored in the `UniqueIterationList`

As the `updateUI` method in `CommissionDetailsPane` is getting lengthy,
let's also abstract out the methods called.

In particular, abstracting the `updateProgressComponent` allows us to call the
same method to refresh the Progress component of the `CommissionDetailsPane`
every time its `UniqueIterationList` is updated (due to adding of a new Iteration,
or deletion of an existing Iteration)- which is needed because the progress status
also depends on the number of Iterations in the `UniqueIterationList`.