Change order of parameters to align with user guide
The order of parameters in the message displayed
after add and edit commands does not align with the user guide.

Ensuring the same order is used in both the app and the UG
could allow user to make cross-references more easily.

Let's update the order of parameters in the app.