Add PropertyModel to ModelManager and Model
ModelManager, which implements the Model interface, only contains
methods to manage the PersonModel class.

however, with the addition of PropertyModel, ModelManager and Model
interface should contain methods that allow other classes to manipulate
the PropertyModel as well.

let's add the same methods and fields that PersonModel has for
PropertyModel in Model and ModelManager.