Fix bugs found in add and delete staff commands
Bugs causes the functionality of the commands to not work as intended
so finding them and resolving them is important to ensure that users
can use them properly and that the team can use them for the
implementation of other components without fear of it failing.

So lets resolve some bugs by adding exceptions for edge cases.