Allow EditCommand to edit homework
The format of EditCommand does not currently match what is planned.
The index of the contact has not yet been removed. It will be
removed in future iterations.

JUnit tests have not been checked yet. Abstraction could possibly
be done better.

Edit functionality for grade progress and attendance will be
implemented similarly.