Fix bug in deleting classes from TA Assist
The deletec command does not unassign students from the class
when it is deleted.
This may cause problems later where there are classes that does
not exist in TA Assist being assigned to students.

Let's change the implementation of deletec command so that it
unassigns students from the classes that will be deleted.