Update Help window UI for Workbook
Add UI scafolding elements for WorkBook commands in JavaFX FXML and CSS
file of Help window.

URL elements in Help windows has id helpMessage and helpMessageContainer which is not specific
enough if more UI elements are added.

Let's change the element id to urlMessage and urlMessageContainer to
clearly seperate it from the other ui elements in the Help window.