Initialise Appointment creation
Right now, patients have a medication field attached to them. There
is no proper use of this field, and we should aim to replace it
with "Upcoming Appointment" and "Past Appointment" fields. For this
to be possible, we need to create the relevent Appointment classes.

As such, the following changes are recommended:
- Create abstract Appointment class
- Create enum to differentiate between types of appointments
- Create PastAppointment and UpcomingAppointment classes
- Add parsing instructions for PastAppointments

Currently, PastAppointments aren't tagged to a person, and this will
be fixed in a future commit. Future commits will also address
parsing instructions for UpcomingAppointment.