UI: fix buggy `CommissionDetailsPane`
Fixes the buggy `CommissionDetailsPane`, noticed that the `updateProgressComponents`
method was not working properly (to replicate, go to a commission that is 'not started'
and add an iteration, you'll notice that the status changed to 'completed'- assumming
the first commission of AB is completed).

This is because `CommissionDetailsPane` stores the first commission that the
`ObservableObject` used to initialise it stores (ie: the first commission
in AB).

Because the `updateProgressComponents` method tries to get the
`completionStatusString` from this stored object, it erronously gets the wrong
completionStatusString during an **iterationList update** (not the commission update),
it'd update the completionStatusString erronously (to the first commission).

This commit fixes this bug by storing the `ObservableObject` instead.