Add internship application stage attribute to internship class
We want to keep track of the stage of application we are at so we know what to expect
next in the application.

Let's add a new field to `AddCommand` and `EditCommand` to include `stage` attribute.

This follows the conventional way of adding and editing  an internship.
As internship application is a process, allowing edits to the stage is
justifiable.

Let's also edit the JUnit tests for these respective classes
to reflect the changes in the main code. Respective JSON files
used in testcases are updated as well.

However, not advised to follow this chunky commit method.