Add parsing logic to FindCommandParser
Find command searches only by name.

A find command that gives the option to search by tags is more
user-friendly and gives the user more freedom, because users might want
to look up groups of users based on a certain user defined criteria.

Let's
* change the FindCommandParser to split command arguments by
CLISyntax prefixes into an ArgumentMultimap
* give ArgumentMultimap the functionality to test if there is only one
prefix stored, and return it
* update the find command to use the appropriate predicate to filter
the address book based on the search type, which is determined by the
prefix found

Splitting the arguments to the find command by prefixes will make the
code easily extensible in the future, should we expand on the find
command by allowing other search conditions as well