Add edit module command
In order to allow users to not only edit tasks but modules as well,
the edit module functionality is implemented. This allows for
greater flexibility when a particular module was named wrongly,
and skips the steps of module deletion and then creation through
the use of a single edit step.