Add mod mark command
Students could add mods to a batchmate, but not mark is as taken or not.

Adding the mark function could allow them to find batchmates to
seek for advice or form study groups more easily.

Let's add the mod mark function.
(NOTE: this has bug now in terms of saving and reading data from file.)