Fix exit command causing NullPointerException
This bug is relatively insignficant in affecting the program's
usability.

Due to a previous change in CommandSpecific (that removed a null check),
HelpCommand and ExitCommand were both afflicted with this issue.

Hence, a new enum for CommandSpecific was added. This is to specify for
commands that don't have any specificity in which window to show (Exit
and Help currently). The overloaded constructor is also changed to
default to this value.

The switch case in MainWindow is also shifted up. This is not a very
important change as it is unlikely to happen again, but this is to hopefully
prevent a similar issue in the future as we will be able to notice any
exceptions caused by the switch block.