Adds Test Ensuring New Tag Addition on First Witness
Add a test in AddCommandTest.java that ensures a tag will be added upon a new person addition through the add command if the tag does not already exist in the mapping.

Introduced ability to peek that the person's list through a completely protected unmodifiable observable list supported with deep copying.