Add bill field to Guest
Guestbook does not keep track of each guest's bill as payment for
the room is made during booking.

Guests may rack up a bill due to room service or damages in the
room.

Let's abstract behaviours common to a bill into a new class called
Bill, and add a Bill field to the Guest model.