Add JCommander and update EditCommandParser
Introduced a new flag for index as the main parameter functionality was not working. Also, there's no way to differentiate between ParameterException, so checking whether a value exists for a flag is not possible currently. Will need to introduce a custom parameter exception subclass to indicate that the value is missing.