Implement rudimentary GetPastAppointmentCommand
Implemented a GetPastAppointmentCommand that prints out the current
list of past appointments in the command result.

Need to add sorting for appointments when they are added to a Person
object based on their dates.
Also need to create a better GUI to display previous appointments
rather than displaying it in command results.