Add Unlink Command
When provided either `PERSON_INDEX`, `INTERNSHIP_INDEX` or both,
Unlink Command will remove the link between the person and
internship specified.

Exceptions will be thrown if link is not found, or specified
person and internship are found to not be linked to persons or
internships other than each other.

- Updated Link Command Success message to display person and
internship names only rather than all details