Add a `BackCommand`
The commands find and rfind causes the gui to display a filtered list.
For rfind specifically, returning to the unfiltered record list requires multiple steps and
is therefore troublesome.

Let us add a backcommand to allow the user to return to the unfiltered list immediately,
making the CLI faster to use.