Update storage to store status of sort by deadline
Each time the user exits and starts the app, the task list is no
longer sorted when the user adds a new task or edits the deadline.

Users who prefer a sort by deadline order will need to sort on
starting the app each time.

Let's update storage to store the status of sort by deadline such
that the user only needs to sort once when they prefer the sort by
deadline order.