Enhance `rfind` command
User is only able to find records that contains record data
which matches the input keywords.

Let's add more functionality to the `rfind` command by allowing the
use of prefixes to specify the fields to match the keywords to.

Records with a field that matches any keyword in its respective field
will be displayed.

**TODO:**
Add more extensive testcases