Bugfixes for Delete/Edit command for person class
Currently, when a modification is made to the UniquePersonList, and there are events tied to the person, the addressBook is not purged properly.

Let's update the addressBook properly to remove all events associated to people that currently do not exist in the list.

We achieve this by making edits to the existing Delete/Edit command to account for these changes in the UniqueEventList