Update DG mod mark activity diagram
Activity diagram for mod mark contains the PlantUML
loop notation which is not taught or used in CS2103T module.

This should not be used and might be seen as a bug to future reviewers.

Let's update the activity diagram to replace that notation with
other appropriate notations allowed in CS2103T module.