Add ability to sort by criteria, in ascending or descending order
New SortCommand and SortCommandParser classes have been created for this
feature. SortCommandParser takes one prefix as the criteria, and sorts
the SortCommand sorts the students in ascending or descending order. If
more than one prefix is provided, an exception will be thrown.

The Order class was created to follow the convention of having a Prefix
class.

Additionally, the method getFilteredStudentList has been refactored to
getStudentList as the ListView will display be a filtered list or a
sorted list depending on the command. The ListView will also refresh
every time a command is run, as I could not find any other way to
display the sorted list while also switching to a filtered list when
other commands are used.