Add storage for id, sort by id and update other commands with id
Each time the user exits and starts the app, the task list is no
longer sorted by their previous id as a new id is given to each of
them.

Users who prefer a sort by id order will not be able to retain the
initial id order.

Let's update storage to store the id of tasks and to know when to
sort by id.

Additionally, the ids of new tasks will be incorrect since
id count is always initialized to 0 but some task could already
have that id.

Let's also update id to correctly initialize id count.