Allow filtering of tasks by deadline
Users are currently only able to view the list of tasks without being able to modify it based on any criteria.

Filtering tasks by deadline will allow users to quickly identify the specific tasks that they are looking for based on the input deadline.

Moving forward, more functionalities will be added to provide users with greater flexibility and control over the task list.