Refactor compareTo method for Session, add JUnit test to test nextSession for ListCommand
Session was using LocalDateTime to implement compareTo

This however, was buggy since for eg. Wed 08:00 compareTo Sat 08:00 may result in Wed being bigger which is not what we want.

This is due to Wed for example being assigned a Date field that is greater than the Date field assigned to Sat when using the defaulting in Session implementation.

Fixed this by refactoring compareTo to use own logic.