Add marko feature functionaliy
`marko` functionality allows marking an order as paid and/or delivered
with the -p and -d flag respectively.

In order to support flagging in user input, I have added a new
FlagTokenizer class. Both Flag and Prefix extends ArgumentToken. This is
to allow flexibility in the type of argument tokens that can be stored
in ArgumentMultimap.