Add module name and module credit fields
Modified the module class to accomodate module name and module credit
fields. Modified classes such as AddModuleCommand and
EditModuleCommand and their respective parser classes to
accomodate these new fields. The storage was also modifed via the
JsonAdaptedModule to save and process these fields, and the UI was
changed to diaplay these fields.