Currently, InTrack has no way of sorting all Internships by the date and time of their tasks.
Let's add a sort command to help users sort their internship
applications by the date and time of thier tasks.

`sort a` will sort all internships in an ascending manner, with the
internship with the task with the nearest date and time at the top.

`sort d` will sort all the internships in a descending manner, with
the internship with the task with the furthest date and time at the
top.