Fix more test cases
Failing test cases exposed another weakness in the InternshipComparator.
A desirable characteristic of our comparator (especially for testing) is unambiguity, so let's further specify how we want two empty datetimes to be sorted - by lexicographical order of Company name.

Also, we have to fix the `equals` method on UniqueInternshipList.java as
equality should only be checked w.r.t the sorted list and not the
internal list as that is the external facing interface anyway. This
fixes all the remaining test cases that were failing.