Update message displayed for added internships
Let's notify the user if they added a stage that was not prepared
beforehand by us. If it is something related, they can easily edit the
stage to suit our pre-defined list and get some tips.

This encourages flexibility for the user, instead of outwright rejection.

Testcases are not updated as of this commit.