Add help parameter which displays command usage
Currently, information for how to input commands is located
only in the help window and user guide. Both of these are rather
inconvenient to a user as they use a separate window.

In order to streamline the CLI usage process, I propose to
add a new parameter "-h" which, instead of processing the command,
returns the usage string for the command. This is similar to
other CLI programs such as ffmpeg or pip.

To achieve this, the command is scanned for the help prefix.
If the prefix is present, execution is hijacked via throwing an
exception in the parser. This may not be the most appropriate way
to do things, but it avoids creating new constructors for each
command in order to return new instances of them (because of
the interface Parser).