Add toString method to Command Result to facilitate easier debugging
Commands are tested using the assertCommandSuccess abstraction, which
tests if a command is successfully executed by comparing the expected
CommandResult and reuslting Model with the actual CommandResult and
model produced by running the command.

Giving CommandResult a toString method will smoothen out the process
of debugging test cases, since with a toString method, we will be
able to tell the difference between the expected output and the actual
output at a glance, instead of seeing the locations of the two
CommandResult objects, which is rather cryptic and not very helpful.