Fix deletemember bug
The bug arises when the following commands are
executed on TABS, namely when deletemember on the same
person is executed twice.

This is because the variables personGroupArrayList
and editedPerson are changed, while the old copies
of the variables are not stored. This results in
Set#remove(editedPerson) to not work as intended.

The following commands can be used to test the bug fix:

- addgroup g/CS2103T
- addgroup g/CS

- addmember g/CS2103T n/Alex Yeoh
- addmember g/CS n/Alex Yeoh

- deletemember g/CS2103T n/Alex Yeoh
- deletemember g/CS n/Alex Yeoh