Update edit module feature
Modified the edit module feature by taking into account the
implications on exams and tasks tied to this
module when the module code is modified. In general, a successful
change of the module code will result in the module field of tasks
and exams to change accordingly as well.This is done to ensure
tasks and exams have a valid and existing module field.