Add Language tags
Let's make a distinction between misc. tags and Language tags which a
user can use to indicate what languages they should be proficient in for
a certain application.

Considered making a new class to extend from `Tag.java` but decided it
was not worth the effort, so `languageTags` are basically a separate
`Set` of Tags. (and will be rendered differently later)