Rename filter predicates and edit filter command javadocs
Currently, we only have one type of predicate for filtering of
characteristics for each of property and buyer. the predicate checks
whether the buyer or property contains all the given characteristics
that the user is filtering by.

however, our matchprop and matchbuyer command will require a different
predicate that checks whether the buyer or property contains *any* of
the given characteristics.

thus, an initial clean up and refactoring of the files is done to
prevent confusion later on.