Refactor matchprop
currently, matchprop only returns buyers that are within the price range
and matches all the characteristics of the property.

however, this might result in little to no buyers matched, as it is not
easy to satisfy this condition.

thus we loosen the criteria for a buyer to be matched in matchprop. a
buyer is matched if it is within the price range, OR contains at least
one of the property's characteristic.