Update Edit command
Editing an internship requires the user to input an index of the task
they are editing, even when they cannot see the full details of what
they are editing.

This may cause them to edit the wrong internships since they may not
know what exactly they are overwriting.

Incorporating the SelectCommand to let the user view the details first
before editing prevents such mistakes and makes the code more in line
with the AddTask and DeleteTask commands.

Let's make it necessary to select an internship before editing it and
update the relevant tests to reflect these changes.