Add match all conditions ability in MultiFlagFilterBuyersCommand
initially, MultiFlagFilterBuyersCommand will return buyers that satisfy
at least 1 of the specified conditions.

however, the user might only want buyers that satisfy all the given
conditions, especially if the list of buyers is very high.

let's add the ability to use the -all flag to specify whether we want
all the conditions to be present in the filtered buyers.