Fix a bug in filtering by TaskDeadline
Filter function does not allow a date input that is earlier than the current date.

While this may be correct and intuitive for creating tasks, as deadlines should only occur in the future, it is not so for the input argument for filtering tasks by date. This change allows users to filter their tasks with deadlines that are older than the current date.