Fix bug and add tests to Tag contains predicate
Bug occurs for tags with mixed lower and upper case characters
Initially the test method in the predicate creates a new Tag, and
checks for equality with a full upper case or lower case tagName.

However, tagName is not converted to a full upper/lower case. Code
was badly done as well.

Now, the method converts the set of tags into a stream of tags,
before using the IgnoreCase method in StringUtil against each
tagName directly. (Similar as other String based containsPredicates)