Change fuzzy filtering behaviour for characteristics
currently, if -fuzzy is passed in for the filter commands, the filtered
list will only contain buyers/properties that satisfy all the
characteristics, or at least one of the other conditions (e.g. price).

however, each characteristic should be seen as a condition by itself,
and thus the filtered list should also contain buyers/properties that
contain at least 1 but not all of the given characteristics.