Update delete module and edit module feature
The delete module feature was modified in a way such that the
deletion of a module would result in the deletion of tasks and
exams related to it. This feature is also augmented with more
guiding messages for the user, to improve user friendliness.
The command message usage is changed for both edit module
and delete module features.