Prevent sort command from taking more than one parameter
Sort command will inform the user that a sort is performed but not
perform any sorting if either:
* more than one parameter is passed as argument to the command, or
* a string containing a paramter flag (eg n/, i/ or m/) is passed but
contains other noise (eg. sort plan/time is accepted as a valid command
as it contains n/, but no sorting is performed)

Use switch case in SortCommandParser to sanitise inputs and ensure only
correct inputs are accepted.