Refine ui further
We should not show tips when no card is selected. This occurs for
example when `clear` is executed. If not, the last selected tip will
remain.

We should also set the correct panel upon startup instead of having to
wait for the user to resize the window for the event listener to pick up
on the changes.