Resolve issued affecting the correctness of the unit tests in EditMeetingCommandTest.
Previously, the constructor for a new Meeting needs to take in the date and time in
unprocessed format.
Therefore, the previous iterations of the code had to depend heavily on converting
the date and time format between dd-MM-yyyy HHmm and EEEE, dd MMMM yyyy hh:mm a,
such as through the method getNonProcessedDateAndTime()

However, after moving the processing of date and time forward to the parser
when creating a new meeting, all date and time format in calling the constructor
of Meeting will be in dd MMMM yyyy hh:mm a format, effectivrly bypassing
getNonProcessedDateAndTime(). I hope this will make the code and dateTime
format conversion less complicated.

Now, the EditMeetingCommandParser also validates the date and time inputted
by the user and converts it to EEEE, dd MMMM yyyy hh:mm a format
which will be used to replace the existing date and time of the
meeting to be edited.