Reorder the validation flow of user input in CreateMeetingCommandParser::parse
Previously the parser will validate the date before checking whether the
number of information segments provided is correct, which sometimes results in
an error message informing users of invalid date and time format
when the actual problem is due to missing information.

The preferred sequence of validation should be:
checking if there is missing information --> checking for empty information
--> checking if date and time format is valid

The parser will now check whether the number of information segments provided is correct,
before it proceeds to validate the date.