Fix issue where adding two persons with similar names to a meeting results in an error.
Previously the code checks for whether the name input from
the user leads to ambiguity (whether it matches more than
one person from the address book) -- if it does, an error
is thrown.

However, this results in cases where a valid full name leading
to ambiguity when there are multiple similar names. For
example, when there are contacts called 'John' and 'John Doe',
the command 'meet John ...' is expected to fetch the contact
called John, but instead an exception is thrown as both
contacts above are returned as the query result, causing
ambiguity.

The modified code divides the user input into multiple more
specific cases and is able to return the expected output
given the above case. An additional method to check for
an exact match between the user's input name and the
names of contacts is implemented.