Modify exception handling in CreateMeetingCommand.java
Previously when an exception is caught in CreateMeetingCommand.java,
the programme displays a CommandResult with the error message,
which refreshes the user input.

This created some inconvenience for some of our testers,
as the would like to refer to their previous user input
and modify from there, instead of typing down a new input.

Thus, in the new version, when an exception is caught,
a new CommandException containing the error message is
thrown instead, which gets caught further down and an
error message is displayed in the feedback Window,
while the original input remains in the text box (in
red font) instead of getting refreshed.